[APPS] recreate the cdc_com_port_single demo for the PIC32CZ_ca80_curiosity_ultra configuration. [MH3-84614]

+++ apps/cdc_com_port_single/firmware/src/config/pic32cz_ca80_curiosity_ultra/usb/src/usb_device_local.h
@@ -114,8 +114,8 @@ typedef struct
 {
     /* Control transfer state */
     bool  inProgress;
-    uint32_t rxDataCount;
-    uint32_t expectedRxDataCount;
+    uint8_t rxDataCount;
+    uint8_t expectedRxDataCount;
     SYS_MODULE_INDEX handlerIndex;
     uint8_t * rxBuffer;
